(dup):! created base controller for accessibilityTestControllers
